Started Migration of Middle Node of Linked List
